% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{test_exported}
\alias{test_exported}
\title{Exported test function}
\usage{
test_exported()
}
\value{
10 (as scalar numeric)
}
\description{
This function only serves to test mocking within an exported function.
}
